refactor: unify storage interface
